EspPromocion/assets/requirements.txt
Pillow==3.3.1
python-resize-image==1.1.10
requests==2.11.1
